Allow to set a fetcher (and choose to not unpack)
It is not possible to track a target which is not an archive
since the `--unpack` flag of the `nix-preetch-url` command is always
set.

This patch adds the support of the `fetcher` attribute which can take
two values, `fetchurl` or fetchTarball` which is the default value and
preserves the current behavior. However, if the fetcher attribute is
set to `fetchurl`, the `--unpack` flag is not set, and
`builtins.fetchurl` is used by the Nix evaluation.